# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação das API's",
        "description": "Sistema Gerenciamento de Livraria",
        "contact": {
            "name": "Everton Sena",
            "email": "evertonsena2009@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Book store api"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Listar os autores cadastrados",
                "operationId": "33b4ca0e528a8b6298d9942781a0c85a",
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Autor"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Autores"
                ],
                "summary": "Criar um novo autor",
                "operationId": "d998cf74685b2e9609bcd48ab986f678",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Autor"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Exibe os dados de um autor",
                "operationId": "5317eb225dd93ee043c24cfacda9f26d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Autores"
                ],
                "summary": "Atualizar os dados de um autor",
                "operationId": "54818ef139366ae08ef5747f7c9d9a36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Autor"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Autores"
                ],
                "summary": "Excluir um autor",
                "operationId": "c8d794bbc8f06854b38ce720c4395db8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                }
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Listar todos os livros cadastrados",
                "operationId": "28b9632495ac5f6f66b2e52001d510e6",
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Livro"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Livros"
                ],
                "summary": "Cria um novo livro",
                "operationId": "b10f5fe35eefa47e974dd4b333d870b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Livro"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Exibe os dados de um livro",
                "operationId": "e94109dcb4f4d40b58a210b43d94016d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Livros"
                ],
                "summary": "Atualiza um determinado livro",
                "operationId": "7430c1078c3cb5da552b2fd50409f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Livro"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Livros"
                ],
                "summary": "Exclui um determinado livro",
                "operationId": "41c44a46960ad1a921ba5bfbf16ea7b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                }
            }
        },
        "/subjects": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Listar os assuntos cadastrados",
                "operationId": "53e4736f6088ff2d20d447af79810328",
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Assunto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Cria um novo assunto",
                "operationId": "049f7d60059f28dea00c68c36da46ebb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assunto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/subjects/{id}": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Exibe os dados de um assunto",
                "operationId": "37e502cb02c3df8aae1908289421b567",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Atualiza os dados de um determinado assunto",
                "operationId": "e7bb484b95364bfde83417b6777634f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assunto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Exclui um assunto",
                "operationId": "c925e7e31a2d902355b25d774d9bf3cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Autor": {
                "title": "Autor",
                "properties": {
                    "codau": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "type": "string",
                        "example": "João Gonçalves"
                    }
                },
                "type": "object"
            },
            "Livro": {
                "title": "Livro",
                "properties": {
                    "codl": {
                        "type": "integer",
                        "example": "17"
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Titulo de teste"
                    },
                    "editora": {
                        "type": "string",
                        "example": "Editora de teste"
                    },
                    "edicao": {
                        "type": "integer",
                        "example": "2"
                    },
                    "anopublicacao": {
                        "type": "integer",
                        "example": "2023"
                    },
                    "valor": {
                        "type": "number",
                        "example": "20.00"
                    },
                    "autores": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Autor"
                        }
                    },
                    "assuntos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Assunto"
                        }
                    }
                },
                "type": "object"
            },
            "Assunto": {
                "title": "Assunto",
                "properties": {
                    "codas": {
                        "type": "integer",
                        "example": "1"
                    },
                    "descricao": {
                        "type": "string",
                        "example": "assunto xxx do autor yyy"
                    }
                },
                "type": "object"
            }
        }
    }
}